pyproject.toml
[project]
name = "standalone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai==0.193.2",
    "django==5.2.6",
    "python-dotenv==1.1.1",
]
